Fixed Custom Styles for Bootstrap
I used Kyogre Primal's Color palette instead of Regice's for the btn primary, secondary, and warning colors. Also, I fixed a bug in some templates in which the secondary button (such as the "cancel" button) was too hard to read. I removed the "btn-outline-secondary", and replaced it with the "btn-secondary" class.