BUGFIX: Web App Works if using production.py as Settings
The web app will run on a local machine if you use production.py as the settings for the web app in manage.py. However, the configuration that it has is similar to a web app in a development environment. The only change is that it's now using a SQLite database instead of a Postgres database. Nonetheless, I can still deploy the web app on Pythonanywhere.